Update create-asymmetric-key-transact-sql.md | RSA-HSM support in SQL 2022 (#33839)
* Update create-asymmetric-key-transact-sql.md

related to IcM 618554163.  when using Azure Key Vault Managed HSM you can create keys with algorithm RSA-HSM but this is only supported starting in SQL server 2022. this is otherwise not documented anywhere and I had a case where a customer was trying to use it in SQL 2017, which is not supported.

* Update setup-steps-for-extensible-key-management-using-the-azure-key-vault.md

adding a note to say we can only use RSA-HSM_2048 and RSA-HSM_3072 starting in SQL 2022. related to IcM 618554163